bin: add busybox sed + fix libbb bb_get_chunk_from_file
sed wouldn't build/run: the port's bb_get_chunk_from_file had a stale 'int *end'
signature (1.36.1 uses 'size_t *end') AND stored the terminator char in *end
instead of the chunk length. sed does temp[len-1], so it got garbage; the bug
was masked because the only other caller (sort) passes NULL. Fixed the signature
+ store the length; added sed to the aarch64+x86_64 world.

Verified aarch64: 'sed s/Ubix/UBIX/ /etc/motd' substitutes correctly; sort still
builds. 25 recipe uses of sed across the tree now have a working sed.

+++ contrib/busybox/libbb.h
@@ -934,7 +934,7 @@ FILE  *xfopen_stdin(const char *filename);
 FILE  *fopen_for_read(const char *filename);
 FILE  *fopen_for_write(const char *filename);
 char  *xmalloc_fgetline(FILE *fp);
-char  *bb_get_chunk_from_file(FILE *file, int *end);
+char  *bb_get_chunk_from_file(FILE *file, size_t *end);
 char  bb_process_escape_sequence(const char **ptr);
 void   bb_simple_perror_msg_and_die(const char *s) NORETURN;
 FILE  *fopen_or_warn(const char *path, const char *mode);
